+++ src/LiveComponent/src/LiveComponentBundle.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\UX\LiveComponent;
 
+use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Symfony\UX\LiveComponent\DependencyInjection\Compiler\OptionalDependencyPass;
@@ -24,6 +25,7 @@ final class LiveComponentBundle extends Bundle
 {
     public function build(ContainerBuilder $container): void
     {
-        $container->addCompilerPass(new OptionalDependencyPass());
+        // must run before Symfony\Component\Serializer\DependencyInjection\SerializerPass
+        $container->addCompilerPass(new OptionalDependencyPass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, 100);
     }
 }